requirements.txt
aiohttp==3.10.5
beautifulsoup4==4.12.3
lxml==5.3.0
orjson==3.10.7
rich==13.8.1
python-telegram-bot==21.4
python-daemon==3.0.1
fake-useragent==1.5.1
